requirements.txt
pytest==4.6.9
selenium==3.141.0
urllib3==1.24.3
Appium_Python_Client==1.0.2
